requirements.txt

requests
pycryptodome
pillow
pywin32
